Break onfidoApi.test.js up into smaller chunks
