+++ src/components/utils/integration_tests/onfidoApi.selfie.test.js
@@ -1,12 +1,9 @@
+import fs from 'fs'
 import {
-  uploadDocument,
   uploadLivePhoto,
   uploadSnapshot,
   sendMultiframeSelfie,
-  requestChallenges,
-  uploadLiveVideo,
 } from '../onfidoApi'
-import fs from 'fs'
 
 const API_URL = 'https://api.onfido.com'
 let jwtToken = null
@@ -40,39 +37,6 @@ const EXPECTED_EXPIRED_TOKEN_ERROR = {
   },
 }
 
-const DOCUMENT_VALIDATIONS = {
-  detect_document: 'error',
-  detect_cutoff: 'warn',
-  detect_glare: 'warn',
-  detect_blur: 'warn',
-}
-
-const TEST_DOCUMENT_DATA = {
-  sdkMetadata: {},
-  validations: { ...DOCUMENT_VALIDATIONS },
-  side: 'front',
-  type: 'passport',
-}
-
-const TEST_VIDEO_DATA = {
-  challengeData: {
-    challenges: [
-      {
-        query: [7, 8, 0],
-        type: 'recite',
-      },
-      {
-        query: 'turnRight',
-        type: 'movement',
-      },
-    ],
-    id: 'test-challenge-data',
-    switchSeconds: 2021,
-  },
-  language: 'en_US',
-  sdkMetadata: {},
-}
-
 const PATH_TO_RESOURCE_FILES = `${__dirname}/../../../../test/resources/`
 
 const createEmptyFile = (
@@ -90,104 +54,6 @@ const createEmptyFile = (
           testing asynchronous code written with the callback pattern https://jestjs.io/docs/en/asynchronous
           Work to address this will be done in a separate ticket (CX-6016)
  */
-
-describe('API uploadDocument endpoint', () => {
-  beforeEach(async () => {
-    jwtToken = await new Promise((resolve) => getTestJwtToken(resolve))
-  })
-
-  test('uploadDocument returns expected response on successful upload', (done) => {
-    expect.assertions(12)
-    const testFileName = 'passport.jpg'
-    const onSuccessCallback = (response) => {
-      try {
-        expect(response).toHaveProperty('applicant_id')
-        expect(response).toHaveProperty('created_at')
-        expect(response).toHaveProperty('download_href')
-        expect(response).toHaveProperty('href')
-        expect(response).toHaveProperty('file_name', testFileName)
-        expect(response).toHaveProperty('file_size')
-        expect(response).toHaveProperty('file_type', 'jpg')
-        expect(response).toHaveProperty('id')
-        expect(response).toHaveProperty('sdk_warnings')
-        expect(response).toHaveProperty('side', 'front')
-        expect(response).toHaveProperty('type', 'passport')
-        expect(response).toHaveProperty('issuing_country', null)
-        done()
-      } catch (err) {
-        done(err)
-      }
-    }
-    const data = fs.readFileSync(`${PATH_TO_RESOURCE_FILES}${testFileName}`)
-    const testFile = new File([data], testFileName, {
-      type: 'image/jpeg',
-    })
-    const documentData = {
-      file: testFile,
-      ...TEST_DOCUMENT_DATA,
-    }
-    uploadDocument(
-      documentData,
-      API_URL,
-      jwtToken,
-      (response) => onSuccessCallback(response),
-      (error) => done(error)
-    )
-  })
-
-  test('uploadDocument returns an error if request is made with an expired JWT token', (done) => {
-    expect.hasAssertions()
-    const onErrorCallback = (error) => {
-      try {
-        expect(error).toEqual(EXPECTED_EXPIRED_TOKEN_ERROR)
-        done()
-      } catch (err) {
-        done(err)
-      }
-    }
-    const testFileName = 'passport.jpg'
-    const data = fs.readFileSync(`${PATH_TO_RESOURCE_FILES}${testFileName}`)
-    const testFile = new File([data], testFileName, {
-      type: 'image/jpeg',
-    })
-    const documentData = {
-      file: testFile,
-      ...TEST_DOCUMENT_DATA,
-    }
-    uploadDocument(
-      documentData,
-      API_URL,
-      EXPIRED_JWT_TOKEN,
-      () => done(),
-      onErrorCallback
-    )
-  })
-
-  test('uploadDocument returns an error on uploading an empty file', (done) => {
-    expect.assertions(3)
-    const onErrorCallback = (error) => {
-      try {
-        expect(error.status).toBe(422)
-        expect(error.response.error.type).toBe('validation_error')
-        expect(error.response.error.fields).toHaveProperty('file')
-        done()
-      } catch (err) {
-        done(err)
-      }
-    }
-    const documentData = {
-      file: createEmptyFile(),
-      ...TEST_DOCUMENT_DATA,
-    }
-    uploadDocument(
-      documentData,
-      API_URL,
-      jwtToken,
-      (response) => done(response),
-      onErrorCallback
-    )
-  })
-})
 
 describe('API uploadLivePhoto endpoint', () => {
   beforeEach(async () => {
@@ -523,165 +389,6 @@ describe.skip('API sendMultiframeSelfie endpoint', () => {
       (response) => done(response),
       onErrorCallback,
       (eventString) => console.log(eventString)
-    )
-  })
-})
-
-describe('API uploadLiveVideo endpoint', () => {
-  beforeEach(async () => {
-    jwtToken = await new Promise((resolve) => getTestJwtToken(resolve))
-  })
-
-  test('uploadLiveVideo returns expected response on successful upload', (done) => {
-    expect.assertions(9)
-    const testFileName = 'test-video.webm'
-    const onSuccessCallback = (response) => {
-      try {
-        expect(response).toHaveProperty(
-          'challenge',
-          TEST_VIDEO_DATA.challengeData.challenges
-        )
-        expect(response).toHaveProperty('languages', [
-          { source: 'sdk', language_code: 'en_US' },
-        ])
-        expect(response).toHaveProperty('created_at')
-        expect(response).toHaveProperty('download_href')
-        expect(response).toHaveProperty('href')
-        expect(response).toHaveProperty('file_name', 'blob')
-        expect(response).toHaveProperty('file_size')
-        expect(response).toHaveProperty('file_type', 'video/webm')
-        expect(response).toHaveProperty('id')
-        done()
-      } catch (err) {
-        done(err)
-      }
-    }
-    const data = fs.readFileSync(`${PATH_TO_RESOURCE_FILES}${testFileName}`)
-    const testFile = new Blob([data], {
-      type: 'video/webm',
-    })
-    const videoData = {
-      ...TEST_VIDEO_DATA,
-      blob: testFile,
-    }
-    uploadLiveVideo(
-      videoData,
-      API_URL,
-      jwtToken,
-      (response) => onSuccessCallback(response),
-      (error) => done(error)
-    )
-  })
-
-  test('uploadLiveVideo returns an error if request is made with an expired JWT token', (done) => {
-    expect.hasAssertions()
-    const onErrorCallback = (error) => {
-      try {
-        expect(error).toEqual(EXPECTED_EXPIRED_TOKEN_ERROR)
-        done()
-      } catch (err) {
-        done(err)
-      }
-    }
-    const testFileName = 'test-video.webm'
-    const data = fs.readFileSync(`${PATH_TO_RESOURCE_FILES}${testFileName}`)
-    const testFile = new File([data], testFileName, {
-      type: 'video/webm',
-    })
-    const videoData = {
-      blob: testFile,
-      ...TEST_VIDEO_DATA,
-    }
-    uploadLiveVideo(
-      videoData,
-      API_URL,
-      EXPIRED_JWT_TOKEN,
-      () => done(),
-      onErrorCallback
-    )
-  })
-
-  test('uploadLiveVideo returns an error on uploading an empty file', (done) => {
-    expect.assertions(3)
-    const onErrorCallback = (error) => {
-      try {
-        expect(error.status).toBe(422)
-        expect(error.response.error.type).toBe('validation_error')
-        expect(error.response.error.fields).toHaveProperty(
-          'attachment_file_size'
-        )
-        done()
-      } catch (err) {
-        done(err)
-      }
-    }
-    const emptyVideoBlob = new Blob([], { type: 'video/webm' })
-    const videoData = {
-      blob: emptyVideoBlob,
-      ...TEST_VIDEO_DATA,
-    }
-    uploadLiveVideo(
-      videoData,
-      API_URL,
-      jwtToken,
-      (response) => done(response),
-      onErrorCallback
-    )
-  })
-})
-
-describe('API requestChallenges endpoint', () => {
-  beforeEach(async () => {
-    jwtToken = await new Promise((resolve) => getTestJwtToken(resolve))
-  })
-
-  test('requestChallenges returns a random 3-digit number challenge and a face turn challenge', async () => {
-    expect.assertions(2)
-    const onSuccessCallback = (response, resolve) => {
-      const { challenge } = response.data
-      expect(challenge).toHaveLength(2)
-      // Example challenge response (order of challenge types is random):
-      // [{"query": "turnLeft", "type": "movement"}, {"query": [1, 9, 0], "type": "recite"}]
-      const expectedTypeRegex = /(movement|recite)/
-      expect(challenge).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({
-            type: expect.stringMatching(expectedTypeRegex),
-            query: expect.anything(),
-          }),
-          expect.objectContaining({
-            type: expect.stringMatching(expectedTypeRegex),
-            query: expect.anything(),
-          }),
-        ])
-      )
-      resolve()
-    }
-    await new Promise((resolve, reject) =>
-      requestChallenges(
-        API_URL,
-        jwtToken,
-        (response) => onSuccessCallback(response, resolve),
-        (error) => reject(error)
-      )
-    )
-  })
-
-  test('requestChallenges returns an error if request is made with an expired JWT token', (done) => {
-    expect.hasAssertions()
-    const onErrorCallback = (error) => {
-      try {
-        expect(error).toEqual(EXPECTED_EXPIRED_TOKEN_ERROR)
-        done()
-      } catch (err) {
-        done(err)
-      }
-    }
-    requestChallenges(
-      API_URL,
-      EXPIRED_JWT_TOKEN,
-      (response) => response,
-      onErrorCallback
     )
   })
 })
